Move dom3d to its own file
